OLE-7693 : Added fireRules method in DroolsEditorController class
