+++ ole-app/olefs/src/main/java/org/kuali/ole/deliver/keyvalue/OlePlaceRequestCirculationDeskKeyValue.java
@@ -8,6 +8,8 @@ import org.kie.api.runtime.rule.AgendaGroup;
 import org.kuali.ole.OLEConstants;
 import org.kuali.ole.deliver.bo.OleCirculationDesk;
 import org.kuali.ole.deliver.bo.OleCirculationDeskLocation;
+import org.kuali.ole.deliver.controller.checkout.CircUtilController;
+import org.kuali.ole.deliver.controller.drools.DroolsEditorController;
 import org.kuali.ole.deliver.drools.CustomAgendaFilter;
 import org.kuali.ole.deliver.drools.DroolsKieEngine;
 import org.kuali.ole.deliver.form.OLEPlaceRequestForm;
@@ -31,25 +33,6 @@ public class OlePlaceRequestCirculationDeskKeyValue extends UifKeyValuesFinderBa
 
     private BusinessObjectService businessObjectService;
 
-
-    public void fireRules(List<Object> facts, String[] expectedRules, String agendaGroup) {
-        KieSession session = DroolsKieEngine.getInstance().getSession();
-        for (Iterator<Object> iterator = facts.iterator(); iterator.hasNext(); ) {
-            Object fact = iterator.next();
-            session.insert(fact);
-        }
-
-        if (null != expectedRules && expectedRules.length > 0) {
-            session.fireAllRules(new CustomAgendaFilter(expectedRules));
-        } else {
-            Agenda agenda = session.getAgenda();
-            AgendaGroup group = agenda.getAgendaGroup(agendaGroup);
-            group.setFocus();
-            session.fireAllRules();
-        }
-        session.dispose();
-    }
-
     public BusinessObjectService getBusinessObjectService() {
         if(businessObjectService == null){
             this.businessObjectService = KRADServiceLocator.getBusinessObjectService();
@@ -68,7 +51,7 @@ public class OlePlaceRequestCirculationDeskKeyValue extends UifKeyValuesFinderBa
         OLEPlaceRequestForm olePlaceRequestForm = (OLEPlaceRequestForm) viewModel;
         facts.add(olePlaceRequestForm);
         facts.add(olePlaceRequestForm.getOlePatronDocument());
-        fireRules(facts, null, "pickup-location");
+        new DroolsEditorController().fireRules(facts, null, "pickup-location");
         List<String> pickUpLocationList = new ArrayList<>();
         if (StringUtils.isNotBlank(olePlaceRequestForm.getPickUpLocation())) {
             String[] pickUpLocations = olePlaceRequestForm.getPickUpLocation().split(",");
